# fix: Remove invalid collection_name parameter from VectorDBInitializer.search() call