## 🐛 Problem

After PR #156 successfully upgraded Qdrant server to v1.15.1, the vector backend was still failing with a **code-level parameter error**:

```
TypeError: VectorDBInitializer.search() got an unexpected keyword argument 'collection_name'
```

### Root Cause
**File**: `benchmarks/mcp_performance_benchmark.py` line 107  
**Issue**: Calling `qdrant_client.search(collection_name="context_store", ...)` 
**Problem**: The method signature doesn't accept `collection_name` parameter

## 🔍 Analysis

### Method Signature (Correct)
```python
def search(self, query_vector: list, limit: int = 10, filter_dict: Optional[Dict[str, Any]] = None) -> list:
```

### Benchmark Call (Incorrect)
```python
await qdrant_client.search(
    collection_name="context_store",  # ❌ This parameter doesn't exist
    query_vector=[0.1] * 384,
    limit=5,
)
```

### How It Should Work
The `VectorDBInitializer.search()` method gets collection name from internal config:
```python
collection_name = self.config.get("qdrant", {}).get("collection_name", "project_context")
```

## ✅ Solution

**Remove the invalid `collection_name` parameter from the benchmark call.**

### Before (Broken)
```python
await qdrant_client.search(
    collection_name="context_store",  # ❌ Invalid parameter
    query_vector=[0.1] * 384,
    limit=5,
)
```

### After (Fixed)
```python
await qdrant_client.search(
    query_vector=[0.1] * 384,  # ✅ Correct signature
    limit=5,
)
```

## 🎯 Impact

This fixes the final piece of the vector backend issue:

### Before This Fix
- ✅ Infrastructure: Qdrant v1.15.1 server compatibility (PR #156)
- ❌ **Code**: Method signature mismatch causing TypeError
- ❌ Vector backend still showing 0.0ms timing
- ❌ Vector backend excluded from searches

### After This Fix
- ✅ Infrastructure: Qdrant v1.15.1 server compatibility  
- ✅ **Code**: Method signature matches calls
- ✅ Vector backend timing should show >0.0ms
- ✅ Vector backend participates in searches

## 🧪 Verification

The fix ensures:
1. **No TypeError**: Method called with correct parameters
2. **Benchmark runs**: Performance testing works
3. **Vector backend active**: Should show in `backends_used` array  
4. **Proper timing**: Vector backend >0.0ms instead of 0.0ms

## 📋 Files Changed

- `benchmarks/mcp_performance_benchmark.py`: Remove invalid `collection_name` parameter

**This completes the vector backend fix sequence:**
1. **PR #156**: Infrastructure upgrade (Qdrant v1.9.6 → v1.15.1)
2. **This PR**: Code fix (remove invalid parameter)

🤖 Generated with [Claude Code](https://claude.ai/code)